+++ ProcessUsage.h
@@ -1,4 +1,4 @@
 #include <map>
 #include <windows.h>
 
-int EnumProcessUsage(std::multimap<float, DWORD> &CAN, bool All);
+size_t EnumProcessUsage(std::multimap<float, DWORD> &CAN, bool All);
